networking_calico: Leverage pre-live-migration
During pre-live-migration, Nova hints to us via the binding
that we *will* be moving in the future. The tap, though, is
not setup until the VM has moved.

Thus, we are free and clear to create the WorkloadEndpoint
early and e.g., allow `calico-dhcp-agent` to provision the
port ahead of when the VM lands at the destination.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>